requirements.txt
Django==4.0
django-environ
dj-database-url
django-bootstrap4
django_boost
Pillow